+++ src/main/java/com/cusonar/ib/domain/UsageRate.java
@@ -14,5 +14,5 @@ public class UsageRate implements Serializable {
 	private static final long serialVersionUID = 8719412859598805835L;
 	
 	private int year;
-	private double usageRate;
+	private float usageRate;
 }
